fix(dart,flutter): skip duplicate prepare

+++ src/cli/tools/dart/utils.ts
@@ -7,26 +7,30 @@ export async function prepareDartHome(
   envSvc: EnvService,
   pathSvc: PathService,
 ): Promise<void> {
-  const rootDart = join(envSvc.rootDir, 'root', '.dart');
   // for root
-  await fs.mkdir(rootDart);
-  await fs.writeFile(
-    join(rootDart, 'dartdev.json'),
-    '{ "firstRun": false, "enabled": false }',
-  );
+  const rootDart = join(envSvc.rootDir, 'root', '.dart');
+  if (!(await pathExists(rootDart, true))) {
+    await fs.mkdir(rootDart);
+    await fs.writeFile(
+      join(rootDart, 'dartdev.json'),
+      '{ "firstRun": false, "enabled": false }',
+    );
+  }
 
   // for user
   const dart = join(envSvc.userHome, '.dart');
-  const dartTool = join(envSvc.userHome, '.dart-tool');
+  if (await pathExists(dart, true)) {
+    const dartTool = join(envSvc.userHome, '.dart-tool');
 
-  await pathSvc.createDir(dart);
-  await pathSvc.createDir(dartTool);
+    await pathSvc.createDir(dart);
+    await pathSvc.createDir(dartTool);
 
-  const dartDev = join(dart, 'dartdev.json');
-  await pathSvc.writeFile(dartDev, '{ "firstRun": false, "enabled": false }');
+    const dartDev = join(dart, 'dartdev.json');
+    await pathSvc.writeFile(dartDev, '{ "firstRun": false, "enabled": false }');
 
-  const dartToolTelemetry = join(dartTool, 'dart-flutter-telemetry.config');
-  await pathSvc.writeFile(dartToolTelemetry, 'reporting=0\n');
+    const dartToolTelemetry = join(dartTool, 'dart-flutter-telemetry.config');
+    await pathSvc.writeFile(dartToolTelemetry, 'reporting=0\n');
+  }
 }
 
 export async function preparePubCache(
